Get the remaining 2/5 MOR tests to run correctly.
1. Argument order in Tpetra_Import constructor was reversed; Epetra_Import and
Tpetra_Import have opposite arg source and target order.

2a. ?petra_GatherAllV had been replaced by Teuchos::gatherAll at some point. The
problem is that *petra_GatherAllV is a work-around for a capability that does
not exist in Teuchos, namely a wrapper to MPI_Allgather*v*, (emphasis on the
v). A SolutionCullingStrategy can yield index sets of different sizes on
different ranks, requiring the vector form of the all-gather. I'm restoring the
three work-around files, ?petra_GatherAllV.?pp, to the repo and reverting the
implementation in Albany_SolutionCullingStrategy.cpp. Some find/grep'ing
suggests that only these four tests:
  ./examples/MOR/MOR_TransientHeat2D/input_galerkin_trunc_colloc_exo.xml
  ./examples/MOR/MOR_TransientHeat2D/input_galerkin_trunc_colloc_sample_exo.xml
  ./examples/FELIX_Stokes/FO_Test/input_fo_domeForward.xml
  ./examples/FELIX_Stokes/FO_Test/input_fo_domeAnalysis.xml
depend on these files:
  Albany_SolutionValuesResponseFunction.?pp
  Albany_SolutionCullingStrategy.?pp
  ?petra_GatherAllV.?pp
These files are listed in order of dependence;
Albany_SolutionValuesResponseFunction.?pp is the key file. We might redesign
this bit of code in the future; in particular, we might add an MPI_Allgatherv
wrapper to Teuchos_DefaultMpiComm.hpp and remove our work-arounds.

2b. valgrind reports "Invalid read of size 2" in Epetra_GatherAllV's call to
MPI_Allgatherv. I examined the surrounding code and have concluded that this is
an instance of an issue reported online affecting OpenMPI version <=1.6.

2c. In Tpetra_GatherAllV.hpp, added ifdef ALBANY_64BIT_INT to distinguish
between MPI_INT and MPI_LONG.

3. Also cleaned up two issues in the same code region that weren't causing a
runtime problem: the unused variable ierr (not returned by Tpetra methods), and
use of double rather than GO in a reduceAll.